pyproject.toml
[project]
name = "preplay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.121.2",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "playwright>=1.56.0",
    "pyjwt[crypto]>=2.10.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "ruff>=0.14.5",
]
test = [
    "faker>=39.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-timeout>=2.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (medium speed, stubbed dependencies)",
    "slow: Slow tests using real Playwright browser",
    "e2e: End-to-end tests with full stack",
]

timeout = 300
timeout_method = "thread"

addopts = "-v --strict-markers --tb=short --disable-warnings"
